feat(func): add support for asset read grants
Added a new method `grantReadAssets` to the `LambdaFunction` class that grants the Lambda function read access to the asset bucket configured via props. It throws an error if no asset bucket is configured.

This update allows Lambda functions within the CDK constructs to easily gain the necessary permissions to read assets, which is useful when the Lambda function needs to interact with assets.