feat(func): add support for asset read grants (#284)
* **Please check if the PR fulfills these requirements**
- [X] The commit message describes your change
- [ ] Tests for the changes have been added if possible (for bug fixes / features)
- [X] Docs have been added / updated (for bug fixes / features)


* **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)

Added a new method `grantReadAssets` to the `LambdaFunction` class that grants the Lambda function read access to the asset bucket configured via props. It throws an error if no asset bucket is configured.

This update allows Lambda functions within the CDK constructs to easily gain the necessary permissions to read assets, which is useful when the Lambda function needs to interact with assets.

* **What is the current behavior?** (You can also link to an open issue here)



* **What is the new behavior (if this is a feature change)?**



* **Does this PR introduce a breaking change?** (What changes might users need to make in their setup due to this PR?)

No

* **Other information**: